Move store to engine
- The CustomSocketFactory is used by the env thus has to be available to the env.
Moving it back to engine seems to be the only solution.
